RDCC-233:Code committed to separate internal and external controller logic
