+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/external/ProfessionalUserController.java
@@ -1,0 +1,70 @@
+package uk.gov.hmcts.reform.professionalapi.controller.external;
+
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.Authorization;
+
+import javax.validation.constraints.NotBlank;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.professionalapi.controller.SuperController;
+
+import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsersEntityResponse;
+
+
+@RequestMapping(
+        path = "refdata/external/v1/organisations",
+        produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+)
+@RestController
+@Slf4j
+public class ProfessionalUserController  extends SuperController {
+
+   @ApiOperation(
+            value = "Retrieves the external users with the given organisation",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiParam(
+            name = "showDeleted",
+            type = "string",
+            value = "flag (True/False) to decide deleted users needs to be shown",
+            required = false
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 200,
+                    message = "List of a professional users along with details",
+                    response = ProfessionalUsersEntityResponse.class
+            ),
+            @ApiResponse(
+                    code = 400,
+                    message = "An invalid organisation identifier was provided"
+            ),
+            @ApiResponse(
+                    code = 404,
+                    message = "No organisation was found with the provided organisation identifier"
+            )
+    })
+    @GetMapping(
+            value = "/{orgId}/users",
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    public ResponseEntity<ProfessionalUsersEntityResponse> findUsersByOrganisation(@PathVariable("orgId") @NotBlank String organisationIdentifier,
+                                                                                   @RequestParam(value = "showDeleted", required = false) String showDeleted) {
+
+        log.info("Received request to get users for organisationIdentifier: By External user:ProfessionalUserInternalController" + organisationIdentifier);
+
+        return getProfessionalUsersEntityResponseOrganisationValid(organisationIdentifier, showDeleted);
+    }
+}
