+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/internal/OrganisationInternalController.java
@@ -1,0 +1,243 @@
+package uk.gov.hmcts.reform.professionalapi.controller.internal;
+
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.Authorization;
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+import lombok.NoArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.professionalapi.controller.SuperController;
+import uk.gov.hmcts.reform.professionalapi.controller.request.NewUserCreationRequest;
+import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreationRequest;
+import uk.gov.hmcts.reform.professionalapi.controller.response.OrganisationPbaResponse;
+import uk.gov.hmcts.reform.professionalapi.controller.response.OrganisationResponse;
+import uk.gov.hmcts.reform.professionalapi.controller.response.OrganisationsDetailResponse;
+import uk.gov.hmcts.reform.professionalapi.controller.response.ProfessionalUsersResponse;
+
+
+
+@RequestMapping(
+        path = "refdata/internal/v1/organisations",
+        produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+)
+@RestController
+@Slf4j
+@NoArgsConstructor
+public class OrganisationInternalController extends SuperController {
+
+    // @Secured("SuperUser")
+    @ApiOperation(
+            value = "Creates an Internal Organisation",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 201,
+                    message = "A representation of the created organisation",
+                    response = OrganisationResponse.class
+            )
+    })
+    @PostMapping(
+            consumes = MediaType.APPLICATION_JSON_UTF8_VALUE,
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    @ResponseBody
+    public ResponseEntity<?> createOrganisation(
+            @Valid @NotNull @RequestBody OrganisationCreationRequest organisationCreationRequest) {
+
+        log.info("Received request to create a new organisation for internal users...");
+        return getCreateOrganisation(organisationCreationRequest);
+    }
+
+    //@Secured("OrgAdmin")
+    @ApiOperation(
+            value = "Retrieves organisation details for internal users",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiParam(
+            allowEmptyValue = true,
+            required = true
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 200,
+                    message = "Details of one or more organisations",
+                    response = OrganisationsDetailResponse.class
+            )
+    })
+    @GetMapping(produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<?> retrieveOrganisations(@RequestParam(required = false) String id) {
+
+        return getRetrieveOrganisation(id);
+    }
+
+    @ApiOperation(
+            value = "Retrieves the internal user with the given email address if organisation is active",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiParam(
+            name = "email",
+            type = "string",
+            value = "The email address of the user to return",
+            required = true
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 200,
+                    message = "A representation of a professional user",
+                    response = ProfessionalUsersResponse.class
+            ),
+            @ApiResponse(
+                    code = 400,
+                    message = "An invalid email address was provided"
+            ),
+            @ApiResponse(
+                    code = 404,
+                    message = "No user was found with the provided email address"
+            )
+    })
+    @GetMapping(
+            value = "/users",
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    public ResponseEntity<?> findUserByEmail(@RequestParam(value = "email") String email) {
+
+        return getfindUserByEmail(email);
+    }
+
+    @ApiOperation(
+            value = "Retrieves an organisations payment accounts by super user email for internal user",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 200,
+                    message = "The organisations associated payment accounts",
+                    response = OrganisationPbaResponse.class
+            )
+    })
+    @GetMapping(
+            path = "/pbas",
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    public ResponseEntity<?> retrievePaymentAccountBySuperUserEmail(@NotNull @RequestParam("email") String email) {
+        log.info("Received request to retrieve an organisations payment accounts by email for internal...");
+        return getRetrievePaymentAccountBySuperUserEmail(email);
+    }
+
+    @ApiOperation(
+            value = "Updates an organisation",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            })
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "Updated an organisation"),
+            @ApiResponse(code = 404, message = "If Organisation is not found"),
+            @ApiResponse(code = 400, message = "If Organisation request sent with null/invalid values for mandatory fields")
+    })
+    @PutMapping(
+            value = "/{orgId}",
+            consumes = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    @ResponseBody
+    public ResponseEntity<?> updatesOrganisation(
+            @Valid @NotNull @RequestBody OrganisationCreationRequest organisationCreationRequest,
+            @PathVariable("orgId") @NotBlank String organisationIdentifier) {
+
+        log.info("Received request to update organisation for organisationIdentifier: ");
+        return getUpdateOrganisation(organisationCreationRequest, organisationIdentifier);
+    }
+
+    @ApiOperation(
+            value = "Retrieves the  internal organisation details with the given status ",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiParam(
+            name = "status",
+            type = "string",
+            value = "The organisation details of the status to return",
+            required = true
+
+    )
+
+    @ApiResponses({
+            @ApiResponse(
+                    code = 200,
+                    message = "A representation of a organisation ",
+                    response = OrganisationsDetailResponse.class
+            ),
+            @ApiResponse(
+                    code = 200,
+                    message = "No organisation details found with the provided status "
+            ),
+            @ApiResponse(
+                    code = 400,
+                    message = "Invalid status provided for an organisation"
+            )
+    })
+    @GetMapping(
+            params = {"status"},
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    public ResponseEntity<?> getAllOrganisationDetailsByStatus(@NotNull @RequestParam("status") String status) {
+
+        return  retrieveAllOrganisationDetailsByStatus(status);
+    }
+
+    @ApiOperation(
+            value = "Add an internal user to an organisation",
+            authorizations = {
+                    @Authorization(value = "ServiceAuthorization")
+            }
+    )
+    @ApiResponses({
+            @ApiResponse(
+                    code = 201,
+                    message = "User has been added",
+                    response = OrganisationResponse.class
+            )
+    })
+    @PostMapping(
+            path = "/{orgId}/users/",
+            consumes = MediaType.APPLICATION_JSON_UTF8_VALUE,
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE
+    )
+    @ResponseBody
+    public ResponseEntity<?> addUserToOrganisation(
+            @Valid @NotNull @RequestBody NewUserCreationRequest newUserCreationRequest,
+            @PathVariable("orgId") @NotBlank String organisationIdentifier) {
+
+        log.info("Received request to add a internal new user to an organisation...");
+
+        return addUserToOrganisation(newUserCreationRequest, organisationIdentifier);
+
+    }
+
+}
